# swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "REST API",
    "description": ""
  },
  "host": "localhost:3000",
  "basePath": "/",
  "tags": [],
  "schemes": [
    "http"
  ],
  "consumes": [],
  "produces": [],
  "paths": {
    "/": {
      "get": {
        "tags": [],
        "description": "",
        "parameters": [],
        "responses": {}
      }
    },
    "/paciente/": {
      "post": {
        "tags": [
          "paciente"
        ],
        "description": "Endpoint para cadastrar um paciente.",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "senha": {
                  "example": "any"
                },
                "email": {
                  "example": "any"
                },
                "nome": {
                  "example": "any"
                },
                "cpf": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {}
      }
    },
    "/paciente/login": {
      "post": {
        "tags": [
          "paciente"
        ],
        "description": "Endpoint para login do paciente.",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "senha": {
                  "example": "any"
                },
                "email\r": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {}
      }
    },
    "/paciente/{id}": {
      "get": {
        "tags": [
          "paciente"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {}
      },
      "patch": {
        "tags": [
          "paciente"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {}
      }
    },
    "/paciente/{id}/consultas": {
      "get": {
        "tags": [
          "paciente"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {}
      }
    },
    "/consulta/": {
      "post": {
        "tags": [
          "comsulta"
        ],
        "description": "Endpoint para cadastrar uma consulta.",
        "parameters": [],
        "responses": {}
      }
    },
    "/consulta/{id}": {
      "get": {
        "tags": [
          "comsulta"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {}
      }
    },
    "/consulta/{id}/exames": {
      "get": {
        "tags": [
          "comsulta"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {}
      }
    },
    "/medico/": {
      "post": {
        "tags": [
          "medico"
        ],
        "description": "Endpoint para cadastrar um medico.",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "tipo\r": {
                  "example": "any"
                },
                "senha": {
                  "example": "any"
                },
                "email": {
                  "example": "any"
                },
                "nome": {
                  "example": "any"
                },
                "cpf": {
                  "example": "any"
                },
                "telefone": {
                  "example": "any"
                },
                "endereco": {
                  "example": "any"
                },
                "residenciainicio": {
                  "example": "any"
                },
                "residenciafim": {
                  "example": "any"
                },
                "crm": {
                  "example": "any"
                },
                "titulacao": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {}
      }
    },
    "/medico/login": {
      "post": {
        "tags": [
          "medico"
        ],
        "description": "Endpoint para login do medico.",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "senha": {
                  "example": "any"
                },
                "email\r": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {}
      }
    },
    "/medico/disponivel": {
      "get": {
        "tags": [
          "medico"
        ],
        "description": "",
        "parameters": [],
        "responses": {}
      }
    },
    "/medico/disponivel/{id}": {
      "patch": {
        "tags": [
          "medico"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {}
      }
    },
    "/exame/": {
      "post": {
        "tags": [
          "exame"
        ],
        "description": "",
        "parameters": [],
        "responses": {}
      }
    },
    "/exame/{id}": {
      "post": {
        "tags": [
          "exame"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {}
      }
    },
    "/exame/{id}/realizar": {
      "post": {
        "tags": [
          "exame"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {}
      }
    },
    "/exame/{id}/validar": {
      "patch": {
        "tags": [
          "exame"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {}
      }
    },
    "/laudo/": {
      "post": {
        "tags": [
          "laudo"
        ],
        "description": "",
        "parameters": [],
        "responses": {}
      }
    },
    "/funcionario/": {
      "post": {
        "tags": [
          "funcionario"
        ],
        "description": "Endpoint para cadastrar um funcionario.",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "example": "any"
                },
                "nome": {
                  "example": "any"
                },
                "cpf": {
                  "example": "any"
                },
                "telefone": {
                  "example": "any"
                },
                "endereco": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {}
      }
    },
    "/funcionario/login": {
      "post": {
        "tags": [
          "funcionario"
        ],
        "description": "",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "senha": {
                  "example": "any"
                },
                "email\r": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {}
      }
    },
    "/admin/": {
      "post": {
        "tags": [
          "admin"
        ],
        "description": "Endpoint para cadastrar um administrador.",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "senha": {
                  "example": "any"
                },
                "email": {
                  "example": "any"
                },
                "nome": {
                  "example": "any"
                },
                "cpf": {
                  "example": "any"
                },
                "telefone": {
                  "example": "any"
                },
                "endereco": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {}
      }
    },
    "/admin/login": {
      "post": {
        "tags": [
          "admin"
        ],
        "description": "Endpoint para login do admin.",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "senha": {
                  "example": "any"
                },
                "email\r": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {}
      }
    }
  },
  "definitions": {}
}